r/linux_virtual_machine_scale_set: moving the extensions out to their own test file
